Update Sparkle installer framework
We are currently using a very old version of Sparkle, back when the
sandbox-compatible XPC branch came out. There has been a lot of
work done on Sparkle and on that branch since then, so this should
be updated.

These XPC services and the Sparkle framework have both been set to
use the hardened runtime. This will make it possible for us to use
the Apple notary service.